TrainingDetails: Fix crash caused by deleting skiday

+++ SkiDays/Core/TrainingDetails/ViewModel/TrainingDetailsViewModel.swift
@@ -5,7 +5,9 @@ class TrainingDetailsViewModel: ObservableObject {
     let service = SkiDayService()
 
     func deleteSkiDay(_ skiDay: SkiDay) {
-        service.deleteVideo(skiDay.video)
+        if skiDay.video != "" {
+            service.deleteVideo(skiDay.video)
+        }
         service.deleteSkiDay(skiDay)
     }
 }
